Work in progress new config.
