+++ src/main/java/io/divolte/server/IncomingRequestProcessingPool.java
@@ -17,6 +17,7 @@
 package io.divolte.server;
 
 import io.divolte.record.DefaultEventRecord;
+import io.divolte.server.config.ValidatedConfiguration;
 import io.divolte.server.hdfs.HdfsFlushingPool;
 import io.divolte.server.ip2geo.ExternalDatabaseLookupService;
 import io.divolte.server.ip2geo.LookupService;
