+++ src/main/resources/reference.conf
@@ -19,9 +19,9 @@ divolte {
   server {
     // The host to which the server binds.
     // Set to a specific IP address to selectively listen on that interface.
-    host = 0.0.0.0
+    //host = 0.0.0.0
     // The bind host can be overridden using the DIVOLTE_HOST environment variable.
-    host = ${?DIVOLTE_HOST}
+    //host = ${?DIVOLTE_HOST}
 
     // The port on which the sever listens.
     port = 8290
