+++ src/test/java/io/divolte/server/hdfs/SessionBinningFileStrategyTest.java
@@ -19,7 +19,7 @@ package io.divolte.server.hdfs;
 import static org.junit.Assert.*;
 import io.divolte.server.AvroRecordBuffer;
 import io.divolte.server.DivolteIdentifier;
-import io.divolte.server.ValidatedConfiguration;
+import io.divolte.server.config.ValidatedConfiguration;
 
 import java.io.File;
 import java.io.IOException;
