--- src/test/resources/poly-test.conf
@@ -1,17 +1,0 @@
-root {
-  foo = "Hello, World"
-  duration = 42 days
-
-  props {
-    foo.bar.baz = foo bar baz
-    magic = false
-    awesome = true
-  }
-  
-  poly {
-    type = one
-    poly_duration = 42 seconds
-    number = 13
-    one = "Woohoo"
-  }
-}
